Added new indexer for the mailer
